requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
pyarrow
matplotlib
seaborn
jupyter